+++ test/specs/components/core/header.component.spec.jsx
@@ -2,16 +2,15 @@ var React = require('react/addons');
 var reactTestUtils = React.addons.TestUtils;
 var Header = require('../../../../web/app/components/core/header.component.jsx');
 var menuActions = require('fluxe').getActions(require('../../../../web/app/components/menu/menu.store').storeName);
-var storeHelper = require('../../../store-helper');
+var testHelper = require('../../../test-helper');
 var _ = require('lodash');
+var Link = require('react-router').Link;
 
 describe('header component', function() {
   beforeEach(function() {
-    storeHelper.resetStores('Menu');
+    testHelper.resetStores('Menu');
 
-    this.component = reactTestUtils.renderIntoDocument(
-      <Header />
-    );
+    this.component = testHelper.getRouterComponent(Header);
   });
 
   it('should have correct initial state', function() {
@@ -58,10 +57,10 @@ describe('header component', function() {
     expect(liElements.length).to.equal(expectedMenuData.length);
 
     _.forEach(liElements, function(liElement, key) {
-      var liProps = reactTestUtils.findRenderedDOMComponentWithTag(liElement, 'a').props;
+      var linkProps = reactTestUtils.findRenderedComponentWithType(liElement, Link).props;
 
-      expect(liProps.href).to.equal(expectedMenuData[key].href);
-      expect(liProps.children).to.equal(expectedMenuData[key].display);
+      expect(linkProps.to).to.equal(expectedMenuData[key].href);
+      expect(linkProps.children).to.equal(expectedMenuData[key].display);
     });
   });
 });