+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## master
+
+- refactored some test code to play nicely with [react-router](https://github.com/rackt/react-router)
+- converted to using [react-router](https://github.com/rackt/react-router)
+- fixed issue with browserify-production not uglifying code
+
 ## 0.1.0
 
 - initial commit